LJVIS-71 06-4 Sõidu- ja puhkeaja vorm: bigint fix insertile

+++ DSL/Resql/ljvis/POST/control-forms/drive-rest-form/driver/insert.sql
@@ -126,7 +126,7 @@ INSERT INTO forms.sp_driver_form (sp_driver_form_key,
                                   notes,
                                   created_by)
 VALUES (nextval('forms.seq_sp_driver_form_key'),
-        :compoundFormKey::INTEGER,
+        :compoundFormKey::BIGINT,
         'sp-' || EXTRACT(YEAR FROM CURRENT_DATE) || '-' ||
         LPAD(currval('forms.seq_sp_driver_form_key')::text, 5, '0') || '/1',
         1,
